pyproject.toml
[project]
name = "project-wed"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "mat_it_get_da", email = "ggaebi011@naver.com" }
]
dependencies = [
    "openai>=2.7.1",
    "dotenv>=0.9.9",
    "python-dotenv>=1.2.1",
    "hydra-core>=1.3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/project_wed"]
